Move getAuBrandMomentBanner file into components
